Fixing failed test cases
